rust-server-cleanup(python-package,non_mechanical_provable): move rust server integration package
