+++ python/sglang/srt/rust_server/__init__.py
@@ -1,0 +1,19 @@
+"""Embedded Rust server integration."""
+
+from sglang.srt.rust_server.server import (
+    NATIVE_MM_FAMILIES,
+    NativeMmFamily,
+    NativeMmHost,
+    NativeMmSpec,
+    RustServer,
+    native_mm_family_for,
+)
+
+__all__ = [
+    "NATIVE_MM_FAMILIES",
+    "NativeMmFamily",
+    "NativeMmHost",
+    "NativeMmSpec",
+    "RustServer",
+    "native_mm_family_for",
+]
