+++ test/registered/unit/multimodal/rust/shared/test_partition_cores.py
@@ -12,7 +12,7 @@ from sglang.test.test_utils import CustomTestCase, maybe_stub_sgl_kernel
 
 maybe_stub_sgl_kernel()
 
-from sglang.srt.managers.rust_server import RustServer  # noqa: E402
+from sglang.srt.rust_server import RustServer  # noqa: E402
 
 register_cpu_ci(est_time=1, suite="base-a-test-cpu")
